break out links.ts from metamodel.ts (using an agent)
